* Fix v3_send_audio to set proper pcm_length when resampling with speex dsp. * Make v3_pcmlength_for_rate always return even value. * Fix JNI wrappers for functions above.
We can now send completely random static without getting banned!

git-svn-id: http://svn.mangler.org/mangler/branches/android@636 f654edcb-a5fb-4f28-a0e7-68c327795dae

+++ project/src/org/mangler/VentriloInterface.java
@@ -26,6 +26,6 @@ public class VentriloInterface {
 	public static native int 		channelcount();
 	public static native void		startprocessing();
 	public static native boolean	recv();
-	public static native void		sendaudio(short[] pcm, int size, int rate);
+	public static native void		sendaudio(byte[] pcm, int size, int rate);
 	public static native int		pcmlengthforrate(int rate);
 }
